codemonk_backend/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery[redis]
drf-yasg
django-cors-headers
